feat: agrego mensaje de error al borrar padre

+++ src/app/(loggedin)/parent/removeParent.ts
@@ -2,17 +2,41 @@
 
 import { getCurrentProfilePrismaClient } from "@/lib/prisma_utils";
 
-export async function removeParent(id: number) {
+
+export async function removeParent(id: number): Promise<string | null> {
     try {
         const prisma = await getCurrentProfilePrismaClient();
+        const parent = await prisma.parent.findUnique({
+            where: {
+                id
+            },
+            include: {
+                children: true
+            }
+        });
+        const students = await prisma.student.findMany({
+            where: {
+                parents: {
+                    some: {
+                        id: parent!.id
+                    }
+                }
+            },
+            include: {
+                parents: true
+            }
+        });
+        if (students.some(x => x.parents.length === 1)) {
+            return "No se puede eliminar, hay estudiantes con un solo padre";
+        }
         await prisma.parent.delete({
             where: {
                 id
             }
         });
-        return true;
-    } catch (error) {
+        return null;
+    } catch (error: any) {
         console.error("Error fetching parent:", error);
-        return false;
+        return null;
     }
 }
